Added suspicious process thread check
This checks if a process thread has a start address outside of a legitimate module, in an attempt to find manually mapped dlls. This can be circumvented by using a gadget within a trusted module as the start address such as `jmp rcx`, where the 'context' passed in to thread (as rcx) is the real entrypoint of module.