Constructing custom page tables for easy physical memory access.
Added the constructing of custom page tables for easy physical memory access. In the pml4 we allocate, we copy over all the kernel entries, and take over a usermode pml4e (0-255), and fill it with a pdpt describing 512 huge pages, which allows us to directly access 512gb of system physical memory.